Move statistic helper functions into base Metric class
